refactor: extract dpp route logic to service to align with guidelines
