+++ src/services/dpp_simulator.service.ts
@@ -827,4 +827,158 @@ export class DppSimulatorService {
       },
     });
   }
+
+  public async generateDownloadZip(
+    stockId: string,
+    year: string,
+    skuId: string,
+    cwd: string,
+  ): Promise<{ buffer: Buffer; filename: string }> {
+    let tmpDir: string | null = null;
+    try {
+      const baseOutputsDir = path.join(cwd, "data", stockId, year, "outputs");
+      if (!fs.existsSync(baseOutputsDir)) {
+        throw new Error("Company data not found. Please generate first.");
+      }
+
+      const tmpDirName = `tmp_${skuId}_${Date.now()}`;
+      tmpDir = path.join(baseOutputsDir, tmpDirName);
+      const tmpMockSourcesDir = path.join(tmpDir, "mock_sources");
+      const tmpProductMockDir = path.join(tmpDir, skuId, "mock_sources");
+
+      fs.mkdirSync(tmpMockSourcesDir, { recursive: true });
+      fs.mkdirSync(tmpProductMockDir, { recursive: true });
+
+      const originalBomsPath = path.join(
+        baseOutputsDir,
+        "mock_sources",
+        "boms_and_precursors.json",
+      );
+      const originalGroundTruthPath = path.join(
+        baseOutputsDir,
+        skuId,
+        "mock_sources",
+        `${skuId}_dpp_ground_truth.json`,
+      );
+
+      if (
+        !fs.existsSync(originalBomsPath) ||
+        !fs.existsSync(originalGroundTruthPath)
+      ) {
+        throw new Error("Original mock sources missing.");
+      }
+
+      const tmpBomsPath = path.join(
+        tmpMockSourcesDir,
+        "boms_and_precursors.json",
+      );
+      const tmpGroundTruthPath = path.join(
+        tmpProductMockDir,
+        `${skuId}_dpp_ground_truth.json`,
+      );
+
+      const bomsData = JSON.parse(fs.readFileSync(originalBomsPath, "utf-8"));
+      const groundTruthData = JSON.parse(
+        fs.readFileSync(originalGroundTruthPath, "utf-8"),
+      );
+
+      fs.writeFileSync(tmpBomsPath, JSON.stringify(bomsData, null, 2));
+      fs.writeFileSync(
+        tmpGroundTruthPath,
+        JSON.stringify(groundTruthData, null, 2),
+      );
+
+      const originalSpecsPath = path.join(
+        baseOutputsDir,
+        skuId,
+        "mock_sources",
+        `${skuId}_product_specs.json`,
+      );
+      const tmpSpecsPath = path.join(
+        tmpProductMockDir,
+        `${skuId}_product_specs.json`,
+      );
+      if (fs.existsSync(originalSpecsPath)) {
+        const specsData = JSON.parse(
+          fs.readFileSync(originalSpecsPath, "utf-8"),
+        );
+        fs.writeFileSync(tmpSpecsPath, JSON.stringify(specsData, null, 2));
+      }
+
+      const { convertJsonToCsv } =
+        await import("@/scripts/e2e_seeder/dpp/convert_json_to_csv");
+      const { convertSpecsToCsv } =
+        await import("@/scripts/e2e_seeder/dpp/convert_specs_to_csv");
+
+      await convertJsonToCsv(stockId, year, { baseDirOverride: tmpDir });
+      await convertSpecsToCsv(stockId, year, {
+        baseDirOverride: tmpDir,
+        targetProductId: skuId,
+      });
+
+      const JSZip = (await import("jszip")).default;
+      const zip = new JSZip();
+
+      const generatedCsvPath = path.join(
+        tmpMockSourcesDir,
+        "boms_and_precursors.csv",
+      );
+      if (fs.existsSync(generatedCsvPath)) {
+        zip.file("boms_and_precursors.csv", fs.readFileSync(generatedCsvPath));
+      }
+
+      const generatedSpecsCsvPath = path.join(
+        tmpProductMockDir,
+        `${skuId}_product_specs.csv`,
+      );
+      if (fs.existsSync(generatedSpecsCsvPath)) {
+        zip.file(
+          `${skuId}_product_specs.csv`,
+          fs.readFileSync(generatedSpecsCsvPath),
+        );
+      }
+
+      const blueprintPathSrc = path.join(
+        baseOutputsDir,
+        skuId,
+        "mock_sources",
+        "fastener_blueprint.png",
+      );
+      if (fs.existsSync(blueprintPathSrc)) {
+        zip.file("fastener_blueprint.png", fs.readFileSync(blueprintPathSrc));
+      }
+
+      const compliancePdfPathSrc = path.join(
+        baseOutputsDir,
+        skuId,
+        "system_ingestion",
+        `${skuId}_dpp_compliance_declaration.pdf`,
+      );
+      if (fs.existsSync(compliancePdfPathSrc)) {
+        zip.file(
+          `${skuId}_dpp_compliance_declaration.pdf`,
+          fs.readFileSync(compliancePdfPathSrc),
+        );
+      }
+
+      const zipBuffer = await zip.generateAsync({
+        type: "nodebuffer",
+        compression: "DEFLATE",
+        compressionOptions: { level: 9 },
+      });
+
+      return {
+        buffer: zipBuffer,
+        filename: `DPP_Simulation_${stockId}_${year}_${skuId}.zip`,
+      };
+    } finally {
+      if (tmpDir && fs.existsSync(tmpDir)) {
+        try {
+          fs.rmSync(tmpDir, { recursive: true, force: true });
+        } catch (cleanupErr) {
+          console.error("Cleanup error:", cleanupErr);
+        }
+      }
+    }
+  }
 }
